refactored preprocessing into own subpackage
