+++ cse/lang/PreprocessorStep.py
@@ -1,0 +1,13 @@
+import abc
+
+class PreprocessorStep(object, metaclass=abc.ABCMeta):
+
+
+    @abc.abstractmethod
+    def processAll(self, tokens):
+        return tokens
+
+
+    @abc.abstractmethod
+    def process(self, token):
+        return token
